% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/BlastPhylos.R
\name{add_target}
\alias{add_target}
\title{Add New Target}
\usage{
add_target(SeqDF, newTar, targetName)
}
\arguments{
\item{SeqDF}{Existing sequence dataframe to be appended.}

\item{newTar}{Sequence of new target.}

\item{targetName}{Name of new target species/taxon.}
}
\value{
Returns new sequence dataframe with new target blast data
}
\description{
Add New Target
}
