% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/BlastPhylos.R
\name{align_seqs}
\alias{align_seqs}
\title{Align Sequences}
\usage{
align_seqs(SeqDF)
}
\arguments{
\item{SeqDF}{Dataframe of sequences to be aligned}
}
\value{
Returns aligned sequence in the form of a DNAbin object.
}
\description{
Align Sequences
}
