% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/BlastPhylos.R
\name{blast_seq}
\alias{blast_seq}
\title{Blast Sequence}
\usage{
blast_seq(refseq, blastDB, hits, targetName)
}
\arguments{
\item{refseq}{Reference sequence to be BLAST searched.}

\item{blastDB}{The database to be searched against.}

\item{hits}{Sets limit to length of hits list.}

\item{targetName}{Name of the target species/taxon.}
}
\value{
Returns dataframe of Blast results.
}
\description{
Blast Sequence
}
