% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/BlastPhylos.R
\name{bootstrap_tree}
\alias{bootstrap_tree}
\title{Build Phylogenic Tree}
\usage{
bootstrap_tree(AlignedSeqs)
}
\arguments{
\item{AlignedSeqs}{DNAbin object of aligned sequences.}
}
\value{
None
}
\description{
Build Phylogenic Tree
}
