% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/BlastPhylos.R
\name{convert_phyobject}
\alias{convert_phyobject}
\title{Convert to phyDat format}
\usage{
convert_phyobject(aligned_seqs)
}
\arguments{
\item{aligned_seqs}{DNAbin object of aligned sequences.}
}
\value{
Returns PhyloSeq Object
}
\description{
Convert to phyDat format
}
