% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/BlastPhylos.R
\name{NJ_Tree}
\alias{NJ_Tree}
\title{Neighbour Join Tree}
\usage{
NJ_Tree(AlignedSeqs)
}
\arguments{
\item{AlignedSeqs}{DNAbin object of aligned sequences.}
}
\value{
Phylogenic tree in the form of ggtree object
}
\description{
Neighbour Join Tree
}
