% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/BlastPhylos.R
\name{optimize_db_parsimony}
\alias{optimize_db_parsimony}
\title{Optimize parsimony for distance based methods: NJ, UPGMA}
\usage{
optimize_db_parsimony(phy_object)
}
\arguments{
\item{phy_object}{phyDat format object.}
}
\value{
Returns pml object
}
\description{
Optimize parsimony for distance based methods: NJ, UPGMA
}
