% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/BlastPhylos.R
\name{optimize_likelihood}
\alias{optimize_likelihood}
\title{Model selection for tree optimization}
\usage{
optimize_likelihood(phy_object, pml_object)
}
\arguments{
\item{phy_object}{phyDat format object.}

\item{pml_object}{PML object.}
}
\value{
Returns pml object
}
\description{
Model selection for tree optimization
}
