% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/BlastPhylos.R
\name{to_fasta}
\alias{to_fasta}
\title{Write Fasta}
\usage{
to_fasta(SeqDF, path)
}
\arguments{
\item{SeqDF}{Dataframe of sequences}

\item{path}{Path where fasta file will be saved}
}
\value{
None
}
\description{
Write Fasta
}
