% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/BlastPhylos.R
\name{bootstrap_tree}
\alias{bootstrap_tree}
\title{Bootstrap tree and save as PDF}
\usage{
bootstrap_tree(fitted_model, bs_iterations, path)
}
\arguments{
\item{fitted_model}{PML object.}

\item{bs_iterations}{Integer for bootstrap iterations.}

\item{path}{Path where PDF will be saved.}
}
\value{
None
}
\description{
Bootstrap tree and save as PDF
}
